Update suggestions according to PR review
