+++ packages/claim-cli/src/utils/confirm-send-transaction.ts
@@ -38,7 +38,7 @@ export async function confirmSendTransaction(
 	console.log(
 		`Estimated Network Fee (${estimatedGas} * ${maxFeePerGas} wei) = ${ethers.formatUnits(estimatedFee)} ETH.`,
 	);
-	console.log(`Your Balance: ${ethers.formatUnits(ethBalance)} ETH`);
+	console.log(`Your Balance: ${ethers.formatUnits(ethBalance)} ETH.`);
 	if (estimatedFee > ethBalance) {
 		console.log('Insufficient Balance for the Transaction.');
 		return process.exit(1);
